requirements.txt
decorator==4.0.6
Django==1.8.11
django-admin-bootstrapped==2.5.7
django-bootstrap3==7.0.0
django-extensions==1.6.1
django-filter==0.12.0
django-rest-swagger==0.3.5
djangorestframework==3.3.2
ipython==4.0.1
ipython-genutils==0.1.0
Markdown==2.6.5
MySQL-python==1.2.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pytz==2015.7
PyYAML==3.11
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
